Make `CustomButton` public
